requirements.txt
PyQt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.14.2.2
pyqtdeploy==3.1.1
pyqtdeploy==3.1.1
python-docx==0.8.10



